[GATE-151] Evaluated and redid sql files that had questionable settings in them.
git-svn-id: http://svn.igniterealtime.org/svn/repos/wildfire/trunk/src/plugins/gateway@6448 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/gateway_sqlserver.sql
@@ -1,19 +1,19 @@
 CREATE TABLE gatewayRegistration (
-   registrationID    BIGINT          NOT NULL,
+   registrationID    INTEGER         NOT NULL,
    jid               NVARCHAR(255)   NOT NULL,
    transportType     NVARCHAR(15)    NOT NULL,
    username          NVARCHAR(255)   NOT NULL,
    password          NVARCHAR(255),
    nickname          NVARCHAR(255),
-   registrationDate  BIGINT          NOT NULL,
-   lastLogin         BIGINT,
+   registrationDate  INTEGER         NOT NULL,
+   lastLogin         INTEGER,
    CONSTRAINT gatewayReg_pk PRIMARY KEY (registrationID)
 );
 CREATE INDEX gatewayReg_jid_idx ON gatewayRegistration (jid);
 CREATE INDEX gatewayReg_type_idx ON gatewayRegistration (transportType);
 
 CREATE TABLE gatewayPseudoRoster (
-   registrationID    BIGINT         NOT NULL,
+   registrationID    INTEGER        NOT NULL,
    username          NVARCHAR(255)  NOT NULL,
    nickname          NVARCHAR(255),
    groups            NVARCHAR(255)
